removed uncessary page content
