+++ gatsby-node.js
@@ -3,7 +3,7 @@ const path = require("path");
 
 require("dotenv").config({
   path: `.env.${process.env.NODE_ENV}`,
-})
+});
 
 exports.onCreateNode = ({ node, getNode, actions }) => {
   const { createNodeField } = actions;
