+++ src/components/diaryPreviewCard/diaryPreviewCard.css
@@ -54,3 +54,49 @@
     text-transform: uppercase;
 }
 
+@media screen and (max-width: 480px) {
+    .card {
+        flex-flow: row;
+        height: unset;
+        width: 100%;
+        min-height: 180px;
+        margin: 20px;
+    }
+
+    .preview-img {
+        border-top-left-radius: 10px;
+        border-top-right-radius: 0;
+        border-bottom-left-radius: 10px;
+        width: 50%;
+    }
+
+    .date {
+        left: 0;
+        right: unset;
+        border-top-left-radius: 10px;
+        border-top-right-radius: 0;
+    }
+
+    .card-contents {
+        height: unset;
+        padding: 10px;
+        width: 45%;
+        align-items: center;
+        text-align: center;
+    }
+
+    .card-title {
+        font-size: 1em;
+    }
+
+    .diary-preview {
+        font-size: 0.8em;
+    }
+
+    #read-more {
+        width: 120%;
+        border: transparent;
+        border-radius: 7px;
+    }
+}
+
