--- src/components/heroSection/heroSection.css
@@ -1,32 +1,0 @@
-.hero-container {
-    background-image: url(../../assets/images/hero-image.png);
-    height: 400px;
-    background-repeat: no-repeat;
-    background-size: 100% 500px;
-    display: flex;
-    justify-content: center;
-}
-
-.hero-section {
-    display: flex;
-    flex-flow: column;
-    align-items: center;
-    margin-top: 150px;
-}
-
-.hero-message {
-    font-size: 2.8rem;
-    color: var(--white);
-    margin: 0.3em 0;
-}
-
-@media screen and (max-width: 480px){
-    .hero-container {
-        height: 320px;
-    }
-
-    .hero-message {
-        font-size: 1.7rem;
-        text-align: center;
-    }
-}
