--- src/components/introduction/introduction.css
@@ -1,53 +1,0 @@
-.introduction {
-    display: flex;
-    flex-flow: row wrap;
-    align-items: center;
-    margin: 50px auto;
-    justify-content: space-evenly;
-}
-
-.quote-banner {
-    text-align: center;
-    display: flex;
-    font-size: 15px;
-    flex-flow: column;
-    justify-content: center;
-    align-items: center;
-    border-radius: 10px;
-    max-width: 90%;
-    margin: 10px;
-}
-
-.quote-banner > p {
-    color: var(--white);
-    font-size: 1.7rem;
-}
-
-@media screen and (max-width: 480px){
-    .introduction {
-        flex-flow: column;
-        margin: 20px auto;
-    }
-
-    .introduction-container {
-        width: 100%;
-    }
-
-    #profile {
-        box-shadow: none;
-        margin: 10px;
-    }
-
-    .intro-message {
-        font-size: 1em;
-    }
-
-    .hero-quote {
-        width: 90%;
-        bottom: 0;
-        left: 0;
-        right: 0;
-        box-shadow: none;
-        margin: 0 auto;
-    }
-}
