--- src/components/introduction/introduction.js
@@ -1,17 +1,0 @@
-import React from "react";
-import "./introduction.css";
-
-const Introduction = () => {
-  return (
-    <div id="home" className="introduction">
-      <div className="quote-banner">
-        <p className="quote">
-          "Remember to celebrate milestones as you prepare for the road ahead"
-        </p>
-        <p className="author">- Nelson Mandela</p>
-      </div>
-    </div>
-  );
-};
-
-export default Introduction;
