+++ src/components/navbars/diaryPageNav/style.css
