+++ src/components/navbars/navbar/style.css
@@ -74,17 +74,15 @@ nav {
     .nav-items {
         position: fixed;
         bottom: 0;
-        width: 100%;
         padding: 0;
         font-size: 12px;
         justify-content: center;
         background-color: var(--secondaryColor);
-        margin: 0;
-        right: 0;
-        left: 0;
-        border-radius: 10px 10px 0 0;
         z-index: 9999;
-        flex-flow: row;
+        width: 60px;
+        height: 60px;
+        border-radius: 50%;
+        left: 2%;
     }
 
     .nav-item, .nav-item > a {
