+++ src/components/profiles/profiles.css
@@ -7,8 +7,8 @@
   }
 
 
-.profiles-container {
-    background-color: var(--secondaryColor);
+.profiles-background {
+    background-color: var(--overlay);
     color: var(--white);
     font-size: 16px;
     position: fixed;
@@ -17,13 +17,23 @@
     right: 0;
     bottom: 0;
     margin: auto;
+    width: 100%;
+    height: 100%;
+    z-index: 99999999;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.content {
     width: 300px;
     height: 300px;
-    z-index: 99999999;
+    background-color: var(--secondaryColor);
+    border-radius: 5px;
     animation: slide 1s forwards;
 }
 
-.profiles-container > ul {
+.content > ul {
     display: flex;
     flex-flow: column;
     align-items: center;
