+++ src/pages/404.js
@@ -1,11 +1,11 @@
-import React from 'react';
+import React from "react";
 
 const Index = () => {
   return (
     <div>
       <h1>Page not Found</h1>
     </div>
-  )
-}
+  );
+};
 
 export default Index;
